Bearer token decoder module for Python
This is a direct port of the JavaScript decoder, so may not be quite as
"pythonic" as it could be. While testing, I noticed Python is a bit more
fussy about certain things than Node.js, so this could probably do with
some try..catch blocks to make it more robust.